Revert "openhab: Don't use legacy java types."
This was incomplete anyway, as return objects still used legacy types.

This reverts commit 1acddb365cae764ddf20a68b7833bedef599f0d6.

+++ configs/brick_imu_v2_config.py
@@ -1359,8 +1359,8 @@ com['openhab'] = {
     this.setLinearAccelerationPeriod(cfg.linearAccelerationUpdateInterval);
     this.setGravityVectorPeriod(cfg.gravityVectorUpdateInterval);
     this.setTemperaturePeriod(cfg.temperatureUpdateInterval);
-    this.setSensorConfiguration(cfg.magnetometerRate, cfg.gyroscopeRange, cfg.gyroscopeBandwidth, cfg.accelerometerRange, cfg.accelerometerBandwidth);
-    this.setSensorFusionMode(cfg.sensorFusionMode);
+    this.setSensorConfiguration(cfg.magnetometerRate.shortValue(), cfg.gyroscopeRange.shortValue(), cfg.gyroscopeBandwidth.shortValue(), cfg.accelerometerRange.shortValue(), cfg.accelerometerBandwidth.shortValue());
+    this.setSensorFusionMode(cfg.sensorFusionMode.shortValue());
     """,
     'channels': [{
             'id': 'Acceleration {}'.format(axis.upper()),
